requirements.txt
appdirs==1.4.4
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dash==2.0.0
dash-bootstrap-components==0.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-Compress==1.10.1
ibm-cloud-sdk-core==3.11.3
ibm-watson==5.2.3
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
numpy-ringbuffer==0.2.1
pandas==1.3.0
Pillow==8.3.1
pipenv==2021.5.29
plotly==5.3.1
praat-parselmouth==0.4.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sounddevice==0.4.2
tenacity==8.0.1
urllib3==1.26.6
virtualenv==20.4.7
virtualenv-clone==0.5.4
websocket-client==1.1.0
Werkzeug==2.0.1
python-dotenv==0.19.0